[TEAM15-545] feat(fridge): default entity integated (#89)

+++ fridge-service/fridge-adapters/src/main/java/com/greenbowl/greenbowlserver/fridge/adapter/out/persistance/categoryitem/CategoryItemPersistenceAdapter.java
@@ -10,6 +10,8 @@ import com.greenbowl.greenbowlserver.fridge.domain.wrapper.Category;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
+import javax.persistence.EntityExistsException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -22,37 +24,49 @@ public class CategoryItemPersistenceAdapter implements
     @Override
     public CategoryItem saveCategoryItem(CategoryItem categoryItem) {
         CategoryItemJpaEntity categoryItemJpaEntity = CategoryItemJpaEntity.from(categoryItem);
+
+        if (categoryItemJpaRepository.existsByCategoryDetailAndDeleteYnFalseAndCategory(
+                categoryItem.getCategoryDetail(),
+                categoryItem.getCategory()
+        )){
+            throw new EntityExistsException("이미 존재하는 카테고리 입니다.");
+        }
         categoryItemJpaRepository.save(categoryItemJpaEntity);
 
         return FridgeJpaEntityToDomainMapper.mapToDomainEntity(categoryItemJpaEntity);
     }
 
     @Override
-    public CategoryItem getCategoryItem(Long userId) {
-        CategoryItemJpaEntity categoryItemJpaEntity = categoryItemJpaRepository.findByUserIdAndDeleteYnFalse(userId);
-        return FridgeJpaEntityToDomainMapper.mapToDomainEntity(categoryItemJpaEntity);
+    public List<CategoryItem> getCategoryItemsByIds(List<Long> ids) {
+        List<CategoryItemJpaEntity> categoryItemJpaEntities
+                = categoryItemJpaRepository.findByIdInAndDeleteYnFalse(ids);
+        return categoryItemJpaEntities.stream()
+                .map(FridgeJpaEntityToDomainMapper::mapToDomainEntity)
+                .collect(Collectors.toList());
     }
 
     @Override
     public List<CategoryItem> getCategoryItemsByUserId(Long userId) {
-        List<CategoryItemJpaEntity> categoryItemJpaEntity = categoryItemJpaRepository.findAllByUserIdAndDeleteYnFalse(userId);
+        List<CategoryItemJpaEntity> categoryItemJpaEntity = categoryItemJpaRepository
+                .findAllByUserIdAndDeleteYnFalse(userId);
         return categoryItemJpaEntity.stream()
                 .map(FridgeJpaEntityToDomainMapper::mapToDomainEntity)
                 .collect(Collectors.toList());
     }
 
     @Override
-    public CategoryItem getCategoryItemById(Long categoryId) {
-        CategoryItemJpaEntity categoryItemJpaEntity = categoryItemJpaRepository.findById(categoryId)
-                .orElseThrow(() -> new IllegalArgumentException("카테고리를 찾을 수 없음: " + categoryId));
-        return FridgeJpaEntityToDomainMapper.mapToDomainEntity(categoryItemJpaEntity);
-    }
-
-    @Override
     public List<CategoryItem> getCategoryItemsByUserIdAndSequence(Long userId, Category category) {
+        List<CategoryItemJpaEntity> result = new ArrayList<>();
+
+        List<CategoryItemJpaEntity> defaultCategoryItem = categoryItemJpaRepository
+                .findAllByUserIdNullAndDeleteYnFalseAndIsDefaultTrue();
+
         List<CategoryItemJpaEntity> categoryItemJpaEntities
                 = categoryItemJpaRepository.findAllByUserIdAndCategoryAndDeleteYnFalse(userId, category);
-        return categoryItemJpaEntities.stream()
+        result.addAll(defaultCategoryItem);
+        result.addAll(categoryItemJpaEntities);
+
+        return result.stream()
                 .map(FridgeJpaEntityToDomainMapper::mapToDomainEntity)
                 .collect(Collectors.toList());
     }
@@ -60,7 +74,12 @@ public class CategoryItemPersistenceAdapter implements
     @Override
     public void deleteCategoryItem(Long userId, DeleteCategoryItemCommand deleteCategoryItemCommand) {
         CategoryItemJpaEntity categoryItemJpaEntity
-                = categoryItemJpaRepository.findByUserIdAndIdAndDeleteYnFalse(userId, deleteCategoryItemCommand.getId());
+                = categoryItemJpaRepository.findByUserIdAndIdAndDeleteYnFalse(userId, deleteCategoryItemCommand.getId())
+                .orElseThrow(()-> new IllegalArgumentException("존재하지 않는 항목입니다."));
+
+        if (categoryItemJpaEntity.isDefault()){
+            throw new IllegalArgumentException("기본 항목은 삭제할 수 없습니다.");
+        }
         categoryItemJpaEntity.deleteCategoryItem();
     }
